restructure project / start moving to string formatting
